requirements.txt
# User Interface
dash==2.17.1
dash-bootstrap-components==1.6.0

# Chatbot
langchain==0.2.6
langchain-openai==0.1.11

# Environment variables
python-dotenv==1.0.1

# Web server
gunicorn==22.0.0